query

+++ src/main/java/com/getambush/backend/backend_demo/controllers/PriceRange.java
@@ -1,9 +1,7 @@
 package com.getambush.backend.backend_demo.controllers;
 
-import java.math.BigDecimal;
-
 public record PriceRange(
-        BigDecimal min,
-        BigDecimal max
+        Integer min,
+        Integer max
 ) {
 }
